+++ mlflow/spark.py
@@ -50,7 +50,7 @@ from mlflow.utils import databricks_utils
 from mlflow.utils.model_utils import _get_flavor_configuration_from_uri
 from mlflow.utils.annotations import experimental
 from mlflow.tracking._model_registry import DEFAULT_AWAIT_MAX_SLEEP_SECONDS
-from mlflow.utils.autologging_utils import autologging_integration, safe_patch
+from mlflow.utils.autologging import autologging_integration, safe_patch
 
 
 FLAVOR_NAME = "spark"
