+++ mlflow/xgboost.py
@@ -37,7 +37,7 @@ from mlflow.utils.environment import _mlflow_conda_env
 from mlflow.utils.model_utils import _get_flavor_configuration
 from mlflow.exceptions import MlflowException
 from mlflow.utils.annotations import experimental
-from mlflow.utils.autologging_utils import (
+from mlflow.utils.autologging import (
     autologging_integration,
     safe_patch,
     exception_safe_function,
@@ -53,7 +53,7 @@ from mlflow.utils.autologging_utils import (
 # Pylint doesn't detect objects used in class keyword arguments (e.g., metaclass) and considers
 # `ExceptionSafeAbstractClass` as 'unused-import': https://github.com/PyCQA/pylint/issues/1630
 # To avoid this bug, disable 'unused-import' on this line.
-from mlflow.utils.autologging_utils import (  # pylint: disable=unused-import
+from mlflow.utils.autologging import (  # pylint: disable=unused-import
     ExceptionSafeAbstractClass,
 )
